Changed template to match the Cybersecurity template
The chapters are now in their own folder. There is also a folder
for the appendix material.

Folder called Information which contains abstract, acknowledgements,
list of appendix, list of chapters and personal details. In the list
of chapters, there is a command to choose between semester 1 and 2
implementations.

All personal information only needs to be entered once. This is done
in the file ./Information/PersonalDetails.tex