Added PhD template based on existing template
I made several modifications to the existing template:

- I needed to update the Thesis.cls file to add the option to add
supervisor/s to the thesis. The supervisor/s are added in the
Information/PersonalDetails.tex file.

- Abbreviations are now added to the file in the
Information/Abbreviations.tex. They use the glossary package to be more
uniform in use. I provided a basic use case example.

- I moved the dedication and declaration to the Information folder. The
aim of this is so the student may not need to change the Thesis.tex file.

The purpose of moving the abbreviations, dedication and declaration to the
Information folder is so the student will need limited, if any interaction
with the Thesis.tex file.